refactor: view 반환 컨트롤러, 데이터 반환 컨트롤러 분리
